# Optimize startup time using TCP sockets instead of junixsocket and `tput` instead of jline