--- a/integration/feature/full-run-logs/src/FullRunLogsTests.scala
+++ b/integration/feature/full-run-logs/src/FullRunLogsTests.scala
@@ -32,8 +32,8 @@ object FullRunLogsTests extends UtestIntegrationTestSuite {
       assert("\\[\\d+\\] <h1>hello</h1>".r.matches(res.out))
 
       val expectedErrorRegex =
-        s"""==================================================== run --text hello ================================================
-           |======================================================================================================================
+        s"""========================================== run --text hello ======================================
+           |==================================================================================================
            |[build.mill-<digits>/<digits>] compile
            |[build.mill-<digits>] [info] compiling 1 Scala source to ${tester.workspacePath}/out/mill-build/compile.dest/classes ...
            |[build.mill-<digits>] [info] done compiling
@@ -41,8 +41,8 @@ object FullRunLogsTests extends UtestIntegrationTestSuite {
            |[<digits>] [info] compiling 1 Java source to ${tester.workspacePath}/out/compile.dest/classes ...
            |[<digits>] [info] done compiling
            |[<digits>/<digits>] run
-           |[<digits>/<digits>] ============================================ run --text hello ============================================<digits>s
-           |======================================================================================================================"""
+           |[<digits>/<digits>] ================================== run --text hello ==================================<digits>s
+           |=================================================================================================="""
           .stripMargin
           .replaceAll("(\r\n)|\r", "\n")
           .replace('\\', '/')