--- a/main/server/src/mill/main/server/Server.scala
+++ b/main/server/src/mill/main/server/Server.scala
@@ -48,7 +48,8 @@ abstract class Server[T](
 
         while (
           running && {
-            val serverSocket = bindSocket()
+            val serverSocket = new java.net.ServerSocket(0)
+            os.write.over(serverDir / ServerFiles.socketPort, serverSocket.getLocalPort.toString)
             try
               interruptWithTimeout(() => serverSocket.close(), () => serverSocket.accept()) match {
                 case None => false
@@ -67,17 +68,6 @@ abstract class Server[T](
 
       }.getOrElse(throw new Exception("Mill server process already present"))
     finally exitServer()
-  }
-
-  def bindSocket(): AFUNIXServerSocket = {
-    val socketPath = os.Path(ServerFiles.pipe(serverDir.toString()))
-    os.remove.all(socketPath)
-
-    val relFile = socketPath.relativeTo(os.pwd).toNIO.toFile
-    serverLog("listening on socket " + relFile + " " + os.pwd)
-    // Use relative path because otherwise the full path might be too long for the socket API
-    val addr = AFUNIXSocketAddress.of(relFile)
-    AFUNIXServerSocket.bindOn(addr)
   }
 
   def proxyInputStreamThroughPumper(in: InputStream): PipedInputStream = {