--- a/runner/client/src/mill/runner/client/MillProcessLauncher.java
+++ b/runner/client/src/mill/runner/client/MillProcessLauncher.java
@@ -205,9 +205,13 @@ public class MillProcessLauncher {
   }
 
   static void writeTerminalDims(Path serverDir) throws Exception {
-    Files.write(
-        serverDir.resolve(ServerFiles.terminfo),
-        (getTerminalDim("cols") + " " + getTerminalDim("lines")).getBytes());
+    try {
+      Files.write(
+          serverDir.resolve(ServerFiles.terminfo),
+          (getTerminalDim("cols") + " " + getTerminalDim("lines")).getBytes());
+    } catch (Exception e) {
+      Files.write(serverDir.resolve(ServerFiles.terminfo), "".getBytes());
+    }
   }
 
   public static void runTermInfoThread(Path serverDir) throws Exception {